# tui/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "skene",
    "description": "Backend server for the skene analysis engine.",
    "version": "0.5.1"
  },
  "paths": {
    "/session": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Create Session",
        "operationId": "create_session_session_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "x-skene-directory",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Skene-Directory",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "session"
        ],
        "summary": "List Sessions",
        "operationId": "list_sessions_session_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "x-skene-directory",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Skene-Directory",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  },
                  "title": "Response List Sessions Session Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/{session_id}": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "Get Session",
        "operationId": "get_session_session__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/{session_id}/children": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "List Children",
        "operationId": "list_children_session__session_id__children_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  },
                  "title": "Response List Children Session  Session Id  Children Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/{session_id}/message": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "List Messages",
        "operationId": "list_messages_session__session_id__message_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageWithParts"
                  },
                  "title": "Response List Messages Session  Session Id  Message Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Prompt",
        "description": "Record the user message and start the run; progress streams on ``/event``.",
        "operationId": "prompt_session__session_id__message_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/UserMessage"
                    },
                    {
                      "$ref": "#/components/schemas/AssistantMessage"
                    },
                    {
                      "$ref": "#/components/schemas/SyntheticMessage"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "role",
                    "mapping": {
                      "user": "#/components/schemas/UserMessage",
                      "assistant": "#/components/schemas/AssistantMessage",
                      "synthetic": "#/components/schemas/SyntheticMessage"
                    }
                  },
                  "title": "Response Prompt Session  Session Id  Message Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/{session_id}/abort": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Abort",
        "operationId": "abort_session__session_id__abort_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Abort Session  Session Id  Abort Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/{session_id}/permissions": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "List Permissions",
        "description": "Every permission request the session has raised, pending first-created first.",
        "operationId": "list_permissions_session__session_id__permissions_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionRequest"
                  },
                  "title": "Response List Permissions Session  Session Id  Permissions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/{session_id}/permissions/{perm_id}": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Answer Permission",
        "description": "Answer a pending ask; the paused tool handler resumes with the verdict.",
        "operationId": "answer_permission_session__session_id__permissions__perm_id__post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "perm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Perm Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionAnswer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionRequest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssistantMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "assistant",
            "enum": [
              "assistant"
            ]
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "provider": {
            "title": "Provider",
            "type": "string",
            "nullable": true
          },
          "model": {
            "title": "Model",
            "type": "string",
            "nullable": true
          },
          "tokens": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenUsage"
              }
            ],
            "nullable": true
          },
          "finish": {
            "title": "Finish",
            "type": "string",
            "nullable": true
          },
          "error": {
            "title": "Error",
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent",
          "created",
          "id",
          "role",
          "sessionId"
        ],
        "title": "AssistantMessage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageWithParts": {
        "properties": {
          "info": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/UserMessage"
              },
              {
                "$ref": "#/components/schemas/AssistantMessage"
              },
              {
                "$ref": "#/components/schemas/SyntheticMessage"
              }
            ],
            "title": "Info",
            "discriminator": {
              "propertyName": "role",
              "mapping": {
                "assistant": "#/components/schemas/AssistantMessage",
                "synthetic": "#/components/schemas/SyntheticMessage",
                "user": "#/components/schemas/UserMessage"
              }
            }
          },
          "parts": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TextPart"
                },
                {
                  "$ref": "#/components/schemas/ReasoningPart"
                },
                {
                  "$ref": "#/components/schemas/ToolPart"
                },
                {
                  "$ref": "#/components/schemas/MilestonePart"
                },
                {
                  "$ref": "#/components/schemas/ArtifactPart"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "artifact": "#/components/schemas/ArtifactPart",
                  "milestone": "#/components/schemas/MilestonePart",
                  "reasoning": "#/components/schemas/ReasoningPart",
                  "text": "#/components/schemas/TextPart",
                  "tool": "#/components/schemas/ToolPart"
                }
              }
            },
            "type": "array",
            "title": "Parts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "info"
        ],
        "title": "MessageWithParts",
        "description": "One message plus its parts, as returned by ``GET /session/{id}/message``."
      },
      "MilestonePart": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "messageId": {
            "type": "string",
            "title": "Messageid"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "milestone",
            "enum": [
              "milestone"
            ]
          },
          "milestone": {
            "$ref": "#/components/schemas/CandidateMilestone"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "messageId",
          "milestone",
          "sessionId",
          "type"
        ],
        "title": "MilestonePart",
        "description": "A candidate milestone emitted by an analysis agent's ``emit_milestone``.\n\nStreams live from subagent runs \u2014 ``milestone.stage_id`` is still\n``None`` at this point; classification happens later in\n``finalize_journey`` and is only visible in the ``journey.yaml``\nartifact, not retroactively on these parts."
      },
      "PermissionAnswer": {
        "properties": {
          "reply": {
            "type": "string",
            "enum": [
              "allow",
              "deny"
            ],
            "title": "Reply"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reply"
        ],
        "title": "PermissionAnswer",
        "description": "Body of ``POST /session/{id}/permissions/{permID}``."
      },
      "PermissionRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "allowed",
              "denied"
            ],
            "title": "Status",
            "default": "pending"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "answered": {
            "title": "Answered",
            "type": "integer",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "created",
          "id",
          "sessionId",
          "title",
          "tool"
        ],
        "title": "PermissionRequest",
        "description": "One pending or answered ask, as persisted and streamed."
      },
      "PromptRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "PromptRequest",
        "description": "Body of ``POST /session/{id}/message``."
      },
...
      "ReasoningPart": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "messageId": {
            "type": "string",
            "title": "Messageid"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "reasoning",
            "enum": [
              "reasoning"
            ]
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "messageId",
          "sessionId",
          "text",
          "type"
        ],
        "title": "ReasoningPart"
      },
...
      "Session": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "projectId": {
            "type": "string",
            "title": "Projectid"
          },
          "parentId": {
            "title": "Parentid",
            "type": "string",
            "nullable": true
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "title": {
            "title": "Title",
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "running",
              "error"
            ],
            "title": "Status",
            "default": "idle"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent",
          "created",
          "id",
          "projectId",
          "updated"
        ],
        "title": "Session",
        "description": "One agent conversation.\n\nSubagent runs are child sessions: ``parent_id`` points at the session\nwhose ``task`` tool spawned them."
      },
      "SessionCreateRequest": {
        "properties": {
          "agent": {
            "type": "string",
            "title": "Agent",
            "default": "skene"
          },
          "parentId": {
            "title": "Parentid",
            "type": "string",
            "nullable": true
          },
          "title": {
            "title": "Title",
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SessionCreateRequest"
      },
      "SyntheticMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "synthetic",
            "enum": [
              "synthetic"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "created",
          "id",
          "role",
          "sessionId"
        ],
        "title": "SyntheticMessage",
        "description": "Server-injected content, e.g. a subagent result surfaced to its parent."
      },
      "TextPart": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "messageId": {
            "type": "string",
            "title": "Messageid"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "text",
            "enum": [
              "text"
            ]
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "messageId",
          "sessionId",
          "text",
          "type"
        ],
        "title": "TextPart"
      },
      "TokenUsage": {
        "properties": {
          "input": {
            "type": "integer",
            "title": "Input",
            "default": 0
          },
          "output": {
            "type": "integer",
            "title": "Output",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TokenUsage"
      },
      "ToolPart": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "messageId": {
            "type": "string",
            "title": "Messageid"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "tool",
            "enum": [
              "tool"
            ]
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "callId": {
            "type": "string",
            "title": "Callid"
          },
          "state": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ToolStatePending"
              },
              {
                "$ref": "#/components/schemas/ToolStateRunning"
              },
              {
                "$ref": "#/components/schemas/ToolStateCompleted"
              },
              {
                "$ref": "#/components/schemas/ToolStateError"
              }
            ],
            "title": "State",
            "discriminator": {
              "propertyName": "status",
              "mapping": {
                "completed": "#/components/schemas/ToolStateCompleted",
                "error": "#/components/schemas/ToolStateError",
                "pending": "#/components/schemas/ToolStatePending",
                "running": "#/components/schemas/ToolStateRunning"
              }
            }
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "callId",
          "id",
          "messageId",
          "sessionId",
          "state",
          "tool",
          "type"
        ],
        "title": "ToolPart"
      },
...
      "UserMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "user",
            "enum": [
              "user"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "created",
          "id",
          "role",
          "sessionId"
        ],
        "title": "UserMessage"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}